+++ src/app/video/analog-cameras/details-analog-camera/details-analog-camera.component.ts
@@ -4,6 +4,7 @@ import { Router } from '@angular/router';
 import { CallOut } from './../../../utilities/callout';
 import { AnalogCameraService } from '../analog-camera.service';
 import { Environment } from 'src/app/app.environment';
+import { DialogService } from 'src/app/confirm-dialog/dialog.service';
 
 @Component({
   selector: 'app-details-analog-camera',
@@ -21,6 +22,7 @@ export class DetailsAnalogCameraComponent implements OnInit {
   constructor(
     private router: Router,
     private analogCameraService: AnalogCameraService,
+    private dialogService: DialogService,
   ) { }
 
   ngOnInit() {
@@ -75,25 +77,31 @@ export class DetailsAnalogCameraComponent implements OnInit {
   }
 
   deleteAnalogCamera(id: number){
-    this.loading = true;
+    this.dialogService.openConfirmDialog().afterClosed().subscribe(
+      res => {
+        if ( res ){
+          this.loading = true;
 
-    this.analogCameraService.deleteAnalogCamera(id).subscribe(
-      (data) => {
-        try {
-          if(data['status'] === 'Analog Camera deleted'){
-            this.loading = false;
-            CallOut.deleted = true;
-            this.router.navigate(["/consult-analog-cameras"]);
-          }
-        } catch (error) {
-          console.log('No logrado')
-        }  
-      },
-      error => {
-        this.loading = false;
-        CallOut.addCallOut('error', 'The Analog Camera has not deleted.', 5000)     
+          this.analogCameraService.deleteAnalogCamera(id).subscribe(
+            (data) => {
+              try {
+                if(data['status'] === 'Analog Camera deleted'){
+                  this.loading = false;
+                  CallOut.deleted = true;
+                  this.router.navigate(["/consult-analog-cameras"]);
+                }
+              } catch (error) {
+                console.log('No logrado')
+              }  
+            },
+            error => {
+              this.loading = false;
+              CallOut.addCallOut('error', 'The Analog Camera has not deleted.', 5000)     
+            }
+          );
+        }
       }
-    );
+    )
   }
 
 }
